Move matrix stacks to the intersector
Even after switching the allocator type, this still gives a 9% improvement to average intersection time due to totally avoiding allocations if the intersector is reused and avoiding the need to copy the matrices when cloning array state.

Intersectors were the only thing that used the matrix stacks in ArrayState, so this doesn't change behaviour, but other classes like ComputeBounds needed modifying, too.
It might have been a good idea for these to have been setting the matrices.
If it was, then now they'll need to keep their own matrix stacks.